requirements.txt
requests==2.24.0
networkx==2.5
Flask==1.1.2
matplotlib==3.3.1
beautifulsoup4==4.9.1
Pillow==7.2.0
gunicorn